[FEAT] : 버튼 잘 동작하도록 수정 (#21)

+++ FakeBithumbAssignment/FakeBithumbAssignment/Screens/Coin/Tabs/CandleStickChartTab/CandleStickChartTabViewController.swift
@@ -14,15 +14,15 @@ class CandleStickChartTabViewController: BaseViewController {
     private let orderCurrency: String = "BTC"
     private var interval: BTCandleStickChartInterval = ._1m
     private let intervalButtons: [IntervalButton] = [
-        IntervalButton(interval: ._1m).then { $0.titleLabel?.text = "1분" },
-        IntervalButton(interval: ._3m).then { $0.titleLabel?.text = "3분" },
-        IntervalButton(interval: ._5m).then { $0.titleLabel?.text = "5분" },
-        IntervalButton(interval: ._10m).then { $0.titleLabel?.text = "10분" },
-        IntervalButton(interval: ._30m).then { $0.titleLabel?.text = "30분" },
-        IntervalButton(interval: ._1h).then { $0.titleLabel?.text = "1시간" },
-        IntervalButton(interval: ._6h).then { $0.titleLabel?.text = "6시간" },
-        IntervalButton(interval: ._12h).then { $0.titleLabel?.text = "12시간" },
-        IntervalButton(interval: ._24h).then { $0.titleLabel?.text = "24시간" },
+        IntervalButton(title: "1분", interval: ._1m),
+        IntervalButton(title: "3분",interval: ._3m),
+        IntervalButton(title: "5분",interval: ._5m),
+        IntervalButton(title: "10분",interval: ._10m),
+        IntervalButton(title: "30분",interval: ._30m),
+        IntervalButton(title: "1시간",interval: ._1h),
+        IntervalButton(title: "6시간",interval: ._6h),
+        IntervalButton(title: "12시간",interval: ._12h),
+        IntervalButton(title: "24시간",interval: ._24h)
     ]
     private var candleSticks: [BTCandleStick] = []
     
@@ -30,17 +30,14 @@ class CandleStickChartTabViewController: BaseViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.view.backgroundColor = UIColor.white
-        self.view.addSubview(candleStickChardView)
-        candleStickChardView.snp.makeConstraints { make in
-            make.edges.equalTo(self.view).inset(UIEdgeInsets(top: 30, left: 30, bottom: 30, right: 30))
-        }
+        self.configUI()
+        self.configButtonTarget()
         Task { await self.fetchInitialData() }
-        Timer.scheduledTimer(timeInterval: 3.0, target: self, selector: #selector(refreshData), userInfo: nil, repeats: true)
-        configUI()
+        Timer.scheduledTimer(timeInterval: 10.0, target: self, selector: #selector(refreshData), userInfo: nil, repeats: true)
     }
     
     override func configUI() {
+        self.view.backgroundColor = UIColor.white
         let buttonStackView = UIStackView(arrangedSubviews: self.intervalButtons).then {
             $0.axis = .horizontal
             $0.alignment = .fill
@@ -54,12 +51,28 @@ class CandleStickChartTabViewController: BaseViewController {
             $0.axis = .vertical
             $0.distribution = .fill
             $0.alignment = .fill
-            $0.spacing = 3
+            $0.spacing = 10
         }
         self.view.addSubview(entireStackView)
         entireStackView.snp.makeConstraints { make in
-            make.edges.equalTo(self.view.safeAreaInsets).inset(UIEdgeInsets(top: 100, left: 10, bottom: 10, right: 10))
+            make.edges.equalTo(self.view).inset(UIEdgeInsets(top: 20, left: 20, bottom: 20, right: 20))
         }
+    }
+    
+    private func configButtonTarget() {
+        self.intervalButtons.forEach {
+            $0.addTarget(self, action: #selector(intervalSelected(_:)), for: .touchUpInside)
+        }
+    }
+    
+    @objc func intervalSelected(_ sender: UIButton) {
+        guard let intervalButton: IntervalButton = sender as? IntervalButton else {
+            return
+        }
+        print(intervalButton)
+        self.candleSticks = []
+        self.interval = intervalButton.interval
+        self.refreshData()
     }
 
     private func fetchInitialData() async {
@@ -73,30 +86,30 @@ class CandleStickChartTabViewController: BaseViewController {
             orderCurrency: self.orderCurrency,
             chartIntervals: self.interval
         )
-        // 코어데이터에 저장된게 없을 경우는 응답된걸 그대로 씀
-        guard !fromCoreData.isEmpty else {
-            let transformed: [BTCandleStick] = self.transform(from: fromAPI)
-            self.setCandleStickToChart(of: transformed)
-            return
-        }
         combineData(coreData: fromCoreData, apiData: fromAPI)
     }
     
     @objc private func refreshData() {
         Task {
             let fromAPI: [BTCandleStickResponse] = await self.btCandleStickApiService.requestCandleStick(
-                of: orderCurrency, interval: interval
+                of: self.orderCurrency,
+                interval: self.interval
             )
                 .sorted { $0.date > $1.date }
             combineData(coreData: self.candleSticks, apiData: fromAPI)
-            self.btCandleStickRepository.saveContext()
         }
     }
     
     private func combineData(coreData: [BTCandleStick], apiData: [BTCandleStickResponse]) {
         // 코어데이터에 저장된게 있을 경우는 API와 코어데이터가 연속된 값인지 확인함
-        let latestFromCoreData: BTCandleStick = coreData[0]
-        let recentDataFilteredFromApi: [BTCandleStickResponse] = apiData.filter { $0.date >= latestFromCoreData.date }
+        guard let latestFromCoreData: BTCandleStick = coreData.last else {
+            // 코어데이터에 저장된게 없을 경우에는 api데이터만 사용
+            self.setCandleStickToChart(of: self.transform(from: apiData))
+            return
+        }
+        let recentDataFilteredFromApi: [BTCandleStickResponse] = apiData.filter {
+            $0.date >= latestFromCoreData.date
+        }
         // 다 최신값이면 DB 다 지워주고 새걸로 채워놓아줌
         if recentDataFilteredFromApi.count == apiData.count {
             coreData.forEach { self.btCandleStickRepository.delete(of: $0) }
@@ -107,7 +120,7 @@ class CandleStickChartTabViewController: BaseViewController {
         // 아니라면 코어데이터 최신값 하나 지워주고 필터링된걸로 다 새로 넣어줌
         self.btCandleStickRepository.delete(of: latestFromCoreData)
         let transformed: [BTCandleStick] = self.transform(from: recentDataFilteredFromApi)
-        let combined: [BTCandleStick] = transformed + coreData[1..<coreData.count]
+        let combined: [BTCandleStick] = coreData[0..<coreData.count-1] + transformed
         setCandleStickToChart(of: combined)
     }
